+++ contao/languages/de/modules.php
@@ -5,11 +5,3 @@
  */
 $GLOBALS['TL_LANG']['MOD']['privacy']        = ['Datenschutz', ''];
 $GLOBALS['TL_LANG']['MOD']['protocols']      = ['Protokoll', 'Hier werden alle datenschutzrelevanten Interaktionen gespeichert.'];
-$GLOBALS['TL_LANG']['MOD']['privacy_opt_in'] = ['Opt-In', 'Hier können Sie Opt-In-E-Mails per E-Mail verschicken.'];
-
-/**
- * Frontend modules
- */
-$GLOBALS['TL_LANG']['FMD']['privacy'] = ['Datenschutz', ''];
-
-$GLOBALS['TL_LANG']['FMD'][\HeimrichHannot\PrivacyBundle\Controller\FrontendModule\ProtocolEntryEditorModuleController::TYPE] = ['Protokolleintragseditor', ''];