+++ contao/languages/it/modules.php
@@ -5,10 +5,3 @@
  */
 $GLOBALS['TL_LANG']['MOD']['privacy']        = ['Privacy', ''];
 $GLOBALS['TL_LANG']['MOD']['protocols']      = ['Logs', 'Qui verranno memorizzate tutte le interazioni rilevanti per la privacy.'];
-$GLOBALS['TL_LANG']['MOD']['privacy_opt_in'] = ['Opt-In', 'Qui potete inviare le E-Mail di Opt-In.'];
-
-/**
- * Frontend modules
- */
-$GLOBALS['TL_LANG']['FMD']['privacy']                                                     = ['Privacy', ''];
-$GLOBALS['TL_LANG']['FMD'][\HeimrichHannot\PrivacyBundle\Controller\FrontendModule\ProtocolEntryEditorModuleController::TYPE] = ['Editor voci di log', ''];