+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 All notable changes to this project will be documented in this file.
 
 ## [2.0.0-DEV] - 2025-01-13
+
+- Removed: frontend opt-in module
+- Removed: backend opt-in module